Map POSIX errors to new exceptions.
The exception classes were created, representing errno values, and divided
in 3 categories. An abstract superclass allows the developer to get the
error message, the process backtrace, and a numeric value representing the
logging priority of the error from the exception object.
The namespace of checked::function is formed by common used functions that
in case of an error throw the right exception.